requirements.txt
boto3
pyyaml
psycopg2
geopandas
SQLAlchemy
logging
coloredlogs
arcgis
moto
